feat!: add Nuxt 3.12 support
