+++ date-io-playground/package.json
@@ -23,7 +23,7 @@
     "@types/moment": "^2.13.0",
     "@types/moment-hijri": "^2.1.0",
     "@types/moment-jalaali": "^0.7.6",
-    "@unocss/nuxt": "^0.58.4",
+    "@unocss/nuxt": "^0.61.3",
     "date-fns": "^2.30.0",
     "date-fns-jalali": "^2.19.0-2",
     "dayjs": "^1.11.10",
@@ -31,10 +31,10 @@
     "moment": "^2.29.4",
     "moment-hijri": "^2.1.2",
     "moment-jalaali": "0.9.2",
-    "nuxt": "^3.11.1",
+    "nuxt": "^3.12.3",
     "sass": "^1.63.6",
     "typescript": "^5.4.3",
-    "vue-tsc": "^2.0.7",
+    "vue-tsc": "^2.0.26",
     "vuetify-nuxt-module": "workspace:*"
   }
 }
