+++ package.json
@@ -65,7 +65,7 @@
     "release": "bumpp && npm publish"
   },
   "dependencies": {
-    "@nuxt/kit": "^3.9.0",
+    "@nuxt/kit": "^3.12.3",
     "defu": "^6.1.4",
     "destr": "^2.0.3",
     "local-pkg": "^0.5.0",
@@ -74,7 +74,7 @@
     "ufo": "^1.5.3",
     "unconfig": "^0.3.11",
     "vite-plugin-vuetify": "^2.0.3",
-    "vuetify": "^3.6.5"
+    "vuetify": "^3.6.12"
   },
   "devDependencies": {
     "@antfu/eslint-config": "^0.43.1",
@@ -86,10 +86,10 @@
     "@iconify-json/carbon": "^1.1.21",
     "@iconify-json/mdi": "^1.1.55",
     "@mdi/js": "^7.3.67",
-    "@nuxt/devtools": "^1.0.8",
+    "@nuxt/devtools": "latest",
     "@nuxt/module-builder": "^0.5.5",
-    "@nuxt/schema": "^3.10.2",
-    "@nuxt/test-utils": "^3.11.0",
+    "@nuxt/schema": "^3.12.3",
+    "@nuxt/test-utils": "^3.13.1",
     "@nuxtjs/i18n": "^8.0.0",
     "@parcel/watcher": "^2.3.0",
     "@types/node": "^18",
@@ -104,7 +104,7 @@
     "typescript": "^5.3.3",
     "vite": "^5.0.12",
     "vitest": "^1.4.0",
-    "vue-tsc": "^2.0.7"
+    "vue-tsc": "^2.0.26"
   },
   "pnpm": {
     "peerDependencyRules": {
@@ -114,7 +114,9 @@
     }
   },
   "resolutions": {
-    "@nuxt/kit": "3.11.1"
+    "@nuxt/kit": "3.12.3",
+    "vite": "5.3.2",
+    "vue": "3.4.31"
   },
   "build": {
     "externals": [
