+++ playground/package.json
@@ -15,15 +15,16 @@
     "@iconify-json/mdi": "^1.1.55",
     "@mdi/js": "^7.3.67",
     "luxon": "^3.4.3",
-    "vuetify": "^3.6.5"
+    "vuetify": "^3.6.12"
   },
   "devDependencies": {
+    "@nuxt/devtools": "latest",
     "@nuxtjs/i18n": "^8.2.0",
-    "@unocss/nuxt": "^0.58.4",
-    "nuxt": "^3.11.1",
+    "@unocss/nuxt": "^0.61.3",
+    "nuxt": "^3.12.3",
     "sass": "^1.63.6",
     "typescript": "^5.4.3",
-    "vue-tsc": "^2.0.7",
+    "vue-tsc": "^2.0.26",
     "vuetify-nuxt-module": "workspace:*"
   }
 }
